requirements.txt
affine==2.3.0
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
cycler==0.10.0
Django==3.0.7
django-crispy-forms==1.9.1
gunicorn==20.0.4
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.1.2
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
snuggs==1.4.7
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
